+++ simter-file-dao-r2dbc/src/main/kotlin/tech/simter/file/impl/dao/r2dbc/FileStoreRepository.kt
@@ -1,0 +1,11 @@
+package tech.simter.file.impl.dao.r2dbc
+
+import org.springframework.data.r2dbc.repository.R2dbcRepository
+import tech.simter.file.impl.dao.r2dbc.po.FileStorePo
+
+/**
+ * The reactive repository.
+ *
+ * @author RJ
+ */
+interface FileStoreRepository : R2dbcRepository<FileStorePo, String>