+++ simter-file-dao-r2dbc/src/test/kotlin/tech/simter/file/impl/dao/r2dbc/CreateMethodImplTest.kt
@@ -2,8 +2,9 @@ package tech.simter.file.impl.dao.r2dbc
 
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.data.r2dbc.core.DatabaseClient
+import org.springframework.boot.test.autoconfigure.data.r2dbc.DataR2dbcTest
+import org.springframework.r2dbc.core.DatabaseClient
+import org.springframework.test.context.junit.jupiter.SpringJUnitConfig
 import reactor.kotlin.test.test
 import tech.simter.file.TABLE_FILE
 import tech.simter.file.core.FileDao
@@ -14,7 +15,8 @@ import tech.simter.file.test.TestHelper.randomFileStore
  *
  * @author RJ
  */
-@SpringBootTest(classes = [UnitTestConfiguration::class])
+@SpringJUnitConfig(UnitTestConfiguration::class)
+@DataR2dbcTest
 class CreateMethodImplTest @Autowired constructor(
   private val client: DatabaseClient,
   private val dao: FileDao
@@ -26,7 +28,7 @@ class CreateMethodImplTest @Autowired constructor(
     dao.create(file).test().expectNext(file.id).verifyComplete()
 
     // verify
-    client.execute("select count(*) from $TABLE_FILE where id = :id")
+    client.sql("select count(*) from $TABLE_FILE where id = :id")
       .bind("id", file.id)
       .map { row -> row[0] as Long }
       .one()