+++ simter-file-dao-r2dbc/src/test/kotlin/tech/simter/file/impl/dao/r2dbc/FindListMethodImplTest.kt
@@ -3,8 +3,9 @@ package tech.simter.file.impl.dao.r2dbc
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.data.r2dbc.core.DatabaseClient
+import org.springframework.boot.test.autoconfigure.data.r2dbc.DataR2dbcTest
+import org.springframework.r2dbc.core.DatabaseClient
+import org.springframework.test.context.junit.jupiter.SpringJUnitConfig
 import reactor.kotlin.test.test
 import tech.simter.file.core.FileDao
 import tech.simter.file.core.ModuleMatcher.Companion.autoModuleMatcher
@@ -21,7 +22,8 @@ import java.util.*
  *
  * @author RJ
  */
-@SpringBootTest(classes = [UnitTestConfiguration::class])
+@SpringJUnitConfig(UnitTestConfiguration::class)
+@DataR2dbcTest
 class FindListMethodImplTest @Autowired constructor(
   private val client: DatabaseClient,
   private val dao: FileDao