+++ simter-file-dao-r2dbc/src/test/kotlin/tech/simter/file/impl/dao/r2dbc/UnitTestConfiguration.kt
@@ -2,7 +2,6 @@ package tech.simter.file.impl.dao.r2dbc
 
 import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Import
 
 /**
  * All configuration for this module.
@@ -10,6 +9,5 @@ import org.springframework.context.annotation.Import
  * @author RJ
  */
 @Configuration
-@Import(ModuleConfiguration::class)
 @ComponentScan("tech.simter")
 class UnitTestConfiguration